// src/app/interfaces/location.ts
export interface Locations {
  name: string
  local_names?: LocalNames
  lat: number
  lon: number
  country: string
  state: string
}

export interface LocalNames {
  la: string
  zh: string
  ko: string
  el: string
  be: string
  pt: string
  bg: string
  ja: string
  lv: string
  he: string
  ar: string
  lt: string
  sr: string
  ru: string
  ka: string
  mk: string
  ce: string
  uk: string
}